+++ croud/__main__.py
@@ -457,7 +457,7 @@ command_tree = {
                         help="The organization name to use.",
                     ),
                     Argument(
-                        "--plan-type", type=int, required=True,
+                        "--plan-type", type=int, required=False,
                         choices=[1, 2, 3, 4, 5, 6],
                         help="The support plan to use for the organization.",
                     ),