+++ croud/organizations/commands.py
@@ -32,10 +32,15 @@ def organizations_create(args: Namespace) -> None:
     """
 
     client = Client.from_args(args)
+    import pdb; pdb.set_trace()
+    if args.plan_type:
+        body = {"name": args.name, "plan_type": args.plan_type}
+    else:
+        body = {"name": args.name}
     data, errors = client.send(
         RequestMethod.POST,
         "/api/v2/organizations/",
-        body={"name": args.name, "plan_type": args.plan_type},
+        body=body
     )
     print_response(
         data=data,